pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-loadout"
dynamic = ["version"]
description = "Install artifacts (skills, rules, agents, commands) into coding agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Nick MacCarthy" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
interactive = ["questionary>=2.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.0",
    "ruff>=0.4",
    "types-PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/nickmaccarthy/loadout"
Issues = "https://github.com/nickmaccarthy/loadout/issues"
Repository = "https://github.com/nickmaccarthy/loadout.git"

[tool.hatch.version]
path = "src/loadout/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/loadout"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
